Silicon/Sophgo: fix CMO bug
Refer to Linux kernel, before executing the dcache.civa or the dcache.iva
instruction, the address must be translated to the virtual address in
any cases. The PAGE_OFFSET=0xff60 0000 0000 0000 (modified for Sv39)
which cause the va is not equal to the pa.

In edk2 phase, only the PCIe DMA operations pay attention to the cache
maintanence operations, the physical address should be equal to the virtual
address, using the dcache.ipa and dcache.cipa is more proper. Refer to
the linux-sophgo (5.19).

Signed-off-by: jingyu.li01 <jingyu.li01@sophgo.com>